Fixed a weird issue where ngnix would crash
Removed unused templates (recreated in Twig)
Changed code formatting
Added cookies that are added by us
Changed array() to [] as we are using PHP 7.3 so no reason to use the old syntax.

Additionally, there are 2 files that needed an edit from mybb to truly stop it from creating cookies as the plugin doesn't seem to stop it entirely.
I will update the readme to include the edits.